Seperate serializers into different files
Seperate serializers for better structure and modularization
